Rename crypt to crypto
