+++ nex-rs/src/crypto/rc4.rs
