+++ nex-rs/src/crypto/result.rs
